+++ R/makeNodePSOCK.R
@@ -33,7 +33,7 @@
 #' the name and the value will be the value of `Sys.getenv()` for that
 #' variable.  Non-existing environment variables will be dropped.
 #' These variables are set using `Sys.setenv()`.
-#' An named element with value `NA_character_` will cause that variable to be
+#' A named element with value `NA_character_` will cause that variable to be
 #' unset, which is done via `Sys.unsetenv()`.
 #' 
 #' @param rscript_libs A character vector of \R library paths that will be
@@ -164,7 +164,7 @@
 #' set, then the default is to use \command{ssh} on Unix-like systems,
 #' including macOS as well as Windows 10.  On older MS Windows versions, which
 #' does not have a built-in \command{ssh} client, the default is to use
-#' (i) \command{plink} from the \href{https://www.putty.org/}{\command{PuTTY}}
+#' (i) \command{plink} from the \href{https://www.chiark.greenend.org.uk/~sgtatham/putty/}{\command{PuTTY}}
 #' project, and then (ii) the \command{ssh} client that is distributed with
 #' RStudio.
 #'