+++ inst/testme/_prologue/030.imports.R
@@ -7,3 +7,11 @@ mdebugf <- parallelly:::mdebugf
 pid_exists <- parallelly:::pid_exists
 isFALSE <- parallelly:::isFALSE
 isNA <- parallelly:::isNA
+
+stopf <- parallelly:::stopf
+warnf <- parallelly:::warnf
+msgf <- parallelly:::msgf
+stealth_sample <- parallelly:::stealth_sample
+supportsMulticoreAndRStudio <- parallelly:::supportsMulticoreAndRStudio
+registerClusterTypes <- parallelly:::registerClusterTypes
+isForkedNode <- parallelly:::isForkedNode